Add files via upload
both India and the rest of the globe. According to estimates, cardiac illnesses account for 28.1% of fatalities. More than 17.6 million fatalities, or a large portion of all deaths in 2016, were caused by it in 2016. Therefore, a system that can predict with exact precision and dependability is required for the appropriate and prompt diagnosis as well as the treatment of such diseases. Numerous academics do extensive research utilising a variety of machine learning algorithms to predict heart illness using different datasets that contain numerous factors that lead to heart attacks. Now it is your turn to do a analysis with the given dataset.

 
Project Output Instructions :
•	Perform data pre-processing operations.
•	As a part of data analysis and visualizations draw all the possible plots to provide essential informations and to derive some meaningful insights.
•	Show your correlation matrix of features according to the datasets.
•	Find out accuracy levels of various machine learning techniques such as Support Vector Machines (SVM), K-Nearest Neighbor (KNN), Decision Trees (DT) , Logistic Regression (LR) and Random Forest (RF).
•	Build your Machine learning model for heart disease detection according to the result.